requirements.txt
beautifulsoup4
cryptography
fastapi[standard]
uvicorn[standard]
python-jose[cryptography]
Flask
Flask_JWT_Extended
httpx
icd10_cm
ollama
praw
prawcore
psycopg2_binary
python_daemon
Requests
schedule
simple_icd_10_cm
urllib3
gunicorn

